More big refactoring
Moved ExtensiveSearch into it's own module.
Added comments in the settings.
Cleanup keymap.
Fixed an issue with getting File object from Include cursor.
Removed some useless settings.
